requirements.txt
asgiref==3.7.2
dateconverter==0.6
Django==5.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
